flict: fix git based versioning
by using a more pythonic way of subprocess.
This will ignore any output of git for non-git releases.

Removed the need for decoding the output by using `universal_newlines`.
Split the command into an argument list, set the working directory and
remove the need for running code in a shell (which is bad practice)

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>