flict: change parameter name to avoid collision
Adding supparses in python seems to imply a certain way of storing
arguments and how these arguments are named. In python 3.9.8 there seems
to be a incompatible change that breaks our naming for
display_compatibility. To avoid this its variable gets renamed similar
to the rest of the sub commands.

Signed-off-by: Jens Erdmann <jens.erdmann@web.de>