Requirements: Change to Compatible Releases
Change dependency release version requirements from fixed ranges to more
flexible compatible release clause introduced in PEP 440.
This reduces flexibility a bit but makes it more predictable and easier
to process.
Dependabot makes it easier to maintain up to date dependencies and this
changes makes it easier for dependabot to process the requirements.

Signed-off-by: Jens Erdmann <jens.erdmann@web.de>