flict-config: safety fallback for HOME
if HOME is not set (and there might be even obscure Unix setups
that do that), the app crashes at startup.
Fix that by providing a non-existing dir.
For users that need to have flict-config without HOME
dir being present still FLICT_USERCONFIG is available.
That one can be used to provide a full qualified path of
a cfg file to be used

Relates to #251

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>